Two-step auth: move Generate new codes into the Backup codes card
